refactor_deposit_to_pages_folder
