+++ packages/cashier/src/containers/__tests__/cashier.spec.js
@@ -38,7 +38,7 @@ jest.mock('@deriv/shared', () => {
 jest.mock('Components/account-prompt-dialog', () => jest.fn(() => 'mockedAccountPromptDialog'));
 jest.mock('Components/error-dialog', () => jest.fn(() => 'mockedErrorDialog'));
 jest.mock('../withdrawal', () => jest.fn(() => 'mockedWithdrawal'));
-jest.mock('../deposit', () => jest.fn(() => 'mockedDeposit'));
+jest.mock('Pages/deposit/deposit', () => jest.fn(() => 'mockedDeposit'));
 
 describe('<Cashier />', () => {
     let history;
