+++ packages/cashier/src/containers/__tests__/on-ramp.spec.js
@@ -18,7 +18,7 @@ jest.mock('@deriv/shared/src/utils/screen/responsive', () => ({
     isMobile: jest.fn(),
 }));
 jest.mock('Components/Error/cashier-locked', () => () => <div>CashierLocked</div>);
-jest.mock('Components/Error/deposit-locked', () => () => <div>DepositLocked</div>);
+jest.mock('Pages/deposit/deposit-locked', () => () => <div>DepositLocked</div>);
 jest.mock('Components/on-ramp-provider-card', () => () => <div>OnRampProviderCard</div>);
 jest.mock('Components/on-ramp-provider-popup', () => () => <div>OnRampProviderPopup</div>);
