+++ packages/cashier/src/containers/index.js
@@ -1,10 +1,9 @@
 import AccountTransfer from './account-transfer.jsx';
 import PaymentAgentTransfer from './payment-agent-transfer.jsx';
 import Cashier from './cashier.jsx';
-import Deposit from './deposit.jsx';
 import Withdrawal from './withdrawal.jsx';
 import PaymentAgent from './payment-agent.jsx';
 import P2PCashier from './p2p-cashier.jsx';
 import OnRamp from './on-ramp.jsx';
 
-export { Cashier, Deposit, Withdrawal, PaymentAgent, AccountTransfer, PaymentAgentTransfer, P2PCashier, OnRamp };
+export { Cashier, Withdrawal, PaymentAgent, AccountTransfer, PaymentAgentTransfer, P2PCashier, OnRamp };
