+++ packages/cashier/src/pages/deposit/crypto-deposit.scss
