+++ packages/cashier/src/pages/deposit/index.js
@@ -1,0 +1,3 @@
+import Deposit from './deposit.jsx';
+
+export { Deposit };
