+++ packages/cashier/src/pages/deposit/onboarding/__tests__/cashier-default-details.spec.js
