+++ packages/cashier/src/pages/deposit/onboarding/__tests__/cashier-default-side-note.spec.js
