+++ packages/cashier/src/pages/deposit/onboarding/cashier-default.scss
